refactor: move roots.py and tools.py to operations folder
