+++ src/pynurbs/operations/tools.py
